requirements.txt
langchain
langchain-community
langchain-huggingface
streamlit
pypdf2 
python-dotenv
google-generativeai
